+++ man/examples/ex_crop.r
@@ -41,6 +41,11 @@ rastByRast <- crop(elev, template)
 plot(elev, col = "gray", legend = FALSE)
 plot(rastByRast, add = TRUE)
 
+### Crop raster by manual extent:
+manualRast <- crop(elev, c(2523700, 2542041, -1718812, -1708812))
+plot(elev, col = "gray", legend = FALSE)
+plot(manualRast, add = TRUE)
+
 ### Crop vector by raster:
 
 # For this example, make the SpatRaster smaller, then crop by this.
@@ -131,6 +136,22 @@ legend("topleft",
     lwd = c(NA, NA, 1),
     col = c("black", "red", "blue"),
     lty = c(NA, NA, "solid"),
+    xpd = NA,
+    bg = "white"
+)
+
+# Crop points vector by numeric vector:
+extent <- c(2523700, 2550000, -1740000, -1710000)
+ptsByNumeric <- crop(dypsis, extent)
+plot(dypsis)
+plot(ext(extent), border = "blue", xpd = NA, lty = "dashed", add = TRUE)
+plot(ptsByNumeric, col = "red", add = TRUE)
+legend("bottomleft",
+    legend = c("Dypsis", "Selected", "Extent"),
+    pch = c(16, 16, NA),
+    lwd = c(NA, NA, 1),
+    col = c("black", "red", "blue"),
+    lty = c(NA, NA, "dashed"),
     xpd = NA,
     bg = "white"
 )