+++ NEWS.md
@@ -1,3 +1,27 @@
+# fasterRaster 8.4.1.3 (2026-09-23)
+
+### New functions and functionality
+o `as.data.frame()` and `as.data.table()` now work on `GRaster`s.  
+o `crop()` can now crop rasters or vectors by an extent supplied as a numeric vector or a `SpatExtent` object.  
+o `crossFreq()` tabulates a table of joint frequencies of cells across two or more `GRaster`s.  
+o `layerCor()` can now compare `GRaster`s using a chi-squared test and Kramer's *V*, or the Kruskal-Wallis *H*, or an "`auto`" option that selects the appropriate test based on raster data types.  
+o `maskNA()` can now also create a mask across layers (versus layer-by-layer).  
+o `na.omit()` assigns `NA` to a cell if any cell in a stack of `GRaster`s has an `NA`.  
+o `prod()` calculates the product across cells of a stack of `GRaster`s.  
+o `writeRaster()` now allows users to forgo appending the raster name to the file name when `byLayer` is `TRUE` if the raster has only a single layer.  
+
+### Potentially code-breaking changes
+o Arithmetical operators (e.g., `log()`, `round()`, etc.) return `GRaster`s with the same name as the input (previously, they had returned rasters with the name of the function that was applied).  
+
+### Bug fixes and issues
+o `[` for `GVector`s now selects the proper geometries.  
+o Hidden function `.copyGRaster()` uses **GRASS** tool `g.copy` to make functions like `mask()` more reliable.  
+o `droplevels()` uses syntax for new version of `data.table` package.  
+o `freq()` tries to self-correct for negative cell counts when integer overflow happens on Windows system (a hack for a **GRASS** bug).  
+o `levels<-` correctly assign levels when the object to the right of `<-` is a `list`.  
+o `simplifyGeom()` does nothing to a points vector (i.e., returns the input).  
+o New versions of example data objects that had been in Madagascar-specific coordinate reference system now in Africa Lambert: `madCoast0`, `madCoast4`, `madRivers`, `madDypsis`, `madPpt`, `madTmax`, and `madTmin`.  
+
 # fasterRaster 8.4.1.2 (2026-04-17)
 ### New functions and functionality
 o `makeGRaster()` and `makeGVector()` are now public and allow users to create `G`-objects from files in **GRASS**. (feature request https://github.com/adamlilith/fasterRaster/issues/137 -- thank you, @JacobusS!)